Prepare for removal of deprecated path argument to pytest_collect_file()
Replace usage of the path argument to the pytest pytest_collect_file
hook with usage of the file_path argument introduced in pytest 7. As
pytest 7 is also our oldest supported version, we can rely on the
availability of the file_path argument.

This change will help to avoid breakage when a future pytest version
drops the deprecated path argument.

Also drops other untested code for pytest < 7 compatibility.